stylecop complaints fixed and use of decimal cast changed to m

+++ tests/Integration/Service.Tests/ParcelsModuleSpecs/WhenAdding.cs
@@ -1,54 +1,57 @@
 ﻿namespace Linn.Stores.Service.Tests.ParcelsModuleSpecs
 {
     using System;
+
     using FluentAssertions;
+
     using Linn.Common.Facade;
     using Linn.Stores.Domain.LinnApps;
     using Linn.Stores.Resources;
+
     using Nancy;
     using Nancy.Testing;
+
     using NSubstitute;
+
     using NUnit.Framework;
 
-    public class WhenAdding: ContextBase
+    public class WhenAdding : ContextBase
     {
         [SetUp]
         public void SetUp()
         {
             var parcel = new Parcel
-            {
-                ParcelNumber = 4,
-                SupplierId = 2,
-                DateCreated = new DateTime(),
-                CarrierId = 4,
-                SupplierInvoiceNo = "Bond, James Bond",
-                ConsignmentNo = "007",
-                CartonCount = 0,
-                PalletCount = 0,
-                Weight = (decimal)00.70,
-                DateReceived = new DateTime(),
-                CheckedById = 123456,
-                Comments = "RSN 212, RSN 118"
-            };
+                         {
+                             ParcelNumber = 4,
+                             SupplierId = 2,
+                             DateCreated = new DateTime(),
+                             CarrierId = 4,
+                             SupplierInvoiceNo = "Bond, James Bond",
+                             ConsignmentNo = "007",
+                             CartonCount = 0,
+                             PalletCount = 0,
+                             Weight = 00.70m,
+                             DateReceived = new DateTime(),
+                             CheckedById = 123456,
+                             Comments = "RSN 212, RSN 118"
+                         };
 
             var parcelResource = new ParcelResource
-                             {
-                                 SupplierId = 2,
-                                 DateCreated = new DateTime().ToString("o"),
-                                 CarrierId = 4,
-                                 SupplierInvoiceNo = "Bond, James Bond",
-                                 ConsignmentNo = "007",
-                                 CartonCount = 0,
-                                 PalletCount = 0,
-                                 Weight = (decimal)00.70,
-                                 DateReceived = new DateTime().ToString("o"),
-                                 CheckedById = 123456,
-                                 Comments = "RSN 212, RSN 118"
-                             };
+                                 {
+                                     SupplierId = 2,
+                                     DateCreated = new DateTime().ToString("o"),
+                                     CarrierId = 4,
+                                     SupplierInvoiceNo = "Bond, James Bond",
+                                     ConsignmentNo = "007",
+                                     CartonCount = 0,
+                                     PalletCount = 0,
+                                     Weight = 00.70m,
+                                     DateReceived = new DateTime().ToString("o"),
+                                     CheckedById = 123456,
+                                     Comments = "RSN 212, RSN 118"
+                                 };
 
-            this.ParcelsFacadeService.Add(Arg.Any<ParcelResource>()) 
-                .Returns(new SuccessResult<Parcel>(parcel));
-
+            this.ParcelsFacadeService.Add(Arg.Any<ParcelResource>()).Returns(new SuccessResult<Parcel>(parcel));
 
             this.Response = this.Browser.Post(
                 "/logistics/parcels",
