+++ tests/Integration/Service.Tests/ParcelsModuleSpecs/WhenGettingById.cs
@@ -1,13 +1,18 @@
 ﻿namespace Linn.Stores.Service.Tests.ParcelsModuleSpecs
 {
     using System;
+
     using FluentAssertions;
+
     using Linn.Common.Facade;
     using Linn.Stores.Domain.LinnApps;
     using Linn.Stores.Resources;
+
     using Nancy;
     using Nancy.Testing;
+
     using NSubstitute;
+
     using NUnit.Framework;
 
     public class WhenGettingById : ContextBase
@@ -16,31 +21,26 @@
         public void SetUp()
         {
             var parcel = new Parcel
-            {
-                ParcelNumber = 1,
-                SupplierId = 2,
-                DateCreated = new DateTime(),
-                CarrierId = 4,
-                SupplierInvoiceNo = "Bond, James Bond",
-                ConsignmentNo = "007",
-                CartonCount = 0,
-                PalletCount = 0,
-                Weight = (decimal)00.70,
-                DateReceived = new DateTime(),
-                CheckedById = 123456,
-                Comments = "RSN 212, RSN 118"
-            };
+                         {
+                             ParcelNumber = 1,
+                             SupplierId = 2,
+                             DateCreated = new DateTime(),
+                             CarrierId = 4,
+                             SupplierInvoiceNo = "Bond, James Bond",
+                             ConsignmentNo = "007",
+                             CartonCount = 0,
+                             PalletCount = 0,
+                             Weight = 00.70m,
+                             DateReceived = new DateTime(),
+                             CheckedById = 123456,
+                             Comments = "RSN 212, RSN 118"
+                         };
 
-            this.ParcelsFacadeService.GetById(Arg.Any<int>()) 
-                .Returns(new SuccessResult<Parcel>(parcel));
-
+            this.ParcelsFacadeService.GetById(Arg.Any<int>()).Returns(new SuccessResult<Parcel>(parcel));
 
             this.Response = this.Browser.Get(
                 "/logistics/parcels/1",
-                with =>
-                {
-                    with.Header("Accept", "application/json");
-                }).Result;
+                with => { with.Header("Accept", "application/json"); }).Result;
         }
 
         [Test]
